feat <libs> :add firebase module
