--- src/firebase/firebase.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-import { FirebaseService } from './firebase.service';
-
-@Module({
-  providers: [FirebaseService],
-  exports: [FirebaseService],
-})
-export class FirebaseModule {}
